refactor(desktop): extract pane workspace component
